Admin Template

+++ database/migrations/2019_12_31_184123_create_discussions_table.php
@@ -17,7 +17,7 @@ class CreateDiscussionsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedBigInteger('post_id');
+            $table->unsignedInteger('post_id');
             $table->foreign('post_id')->references('id')->on('posts');
             $table->text('discussion');
             $table->softDeletes();
